Do not allocate RAM dump buffer
The sdio code allocates a buffer for RAM dumps with a static RAM address
with ioremap(). This is very strange and of course fails on i.MX6SX.

We do not want to (and even can not) debug firmware code, so any such dump
does not make sense for end customers. Disable this feature by returning
NULL from the allocate function.